Modify foriegnkey category vs food

+++ src/main/java/com/spring/dev2chuc/nutritious_food/payload/response/UserProfileDTO.java
@@ -17,7 +17,7 @@ public class UserProfileDTO {
     private Integer height;
     private Integer weight;
     private Integer bodyFat;
-    private Integer exerciseIntensity;
+    private Float exerciseIntensity;
     private Integer lbmIndex;
     private Integer bmrIndex;
     private Integer tdeeIndex;
